Squashed a few VS10 warnings.
git-svn-id: svn://svn.code.sf.net/p/segs/code2@252 a19be42f-918d-4297-8077-ab25d1e4f51f

+++ Projects/CoX/Common/NetStructures/src/Costume.cpp
@@ -59,7 +59,7 @@ CharacterCostume CharacterCostume::NullCostume;
 void Costume::storeCharselParts( BitStream &bs )
 {
     bs.StorePackedBits(1,m_parts.size());
-    for(int costume_part=0; costume_part<m_parts.size();costume_part++)
+    for(size_t costume_part=0; costume_part<m_parts.size();costume_part++)
     {
         m_parts[costume_part].m_type=costume_part;
         m_parts[costume_part].serializeto_charsel(bs);
